pyproject.toml
[tool.poetry]
name = "rh-starter"
version = "0.1.0"
description = "TROIA Technical interview"
authors = ["KricejJanezMartin <jkricej@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openmeteo-requests = "^1.2.0"
requests-cache = "^1.2.0"
retry-requests = "^2.0.0"
numpy = "^1.26.4"
pandas = "^2.2.1"
flask = "^3.0.2"
python-dotenv = "^1.0.1"
mysql-connector-python = "^8.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
